docs/requirements.txt
numpydoc
pydata-sphinx-theme
sphinx_automodapi
